[Feat] 회원 정보 수정 기능 구현

+++ src/main/java/com/be/rebook/domain/members/repository/MembersRepository.java
@@ -3,7 +3,7 @@ package com.be.rebook.domain.members.repository;
 import com.be.rebook.domain.members.entity.Members;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MembersRepository extends JpaRepository<Members, Integer> {
+public interface MembersRepository extends JpaRepository<Members, Long> {
 
     Boolean existsByUsername(String username);
